pyproject.toml
[project]
name = "SCIST-Camp-2026-AI-Courses-4yu"
version = "0.1.0"
description = "LLM + RAG + Discord Bot"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ollama>=0.6.1",
    "openai>=2.17.0",
    "py-cord>=2.7.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "langchain-community",
    "langchain-text-splitters",
    "huggingface_hub",
    "faiss-cpu",
    "unstructured",
    "unstructured[docx,pdf]",
    "pypdf",
    "sentence-transformers"
]